+++ lib/src/middleware/require_auth.dart
@@ -3,24 +3,19 @@ import 'package:angel_framework/angel_framework.dart';
 /// Forces Basic authentication over the requested resource, with the given [realm] name, if no JWT is present.
 ///
 /// [realm] defaults to `'angel_auth'`.
-RequestHandler forceBasicAuth({String realm, String userKey: 'user'}) {
+RequestHandler forceBasicAuth<User>({String realm}) {
   return (RequestContext req, ResponseContext res) async {
-    if (req.properties.containsKey(userKey)) return true;
+    if (req.container.has<User>()) return true;
 
     res
       ..statusCode = 401
       ..headers['www-authenticate'] = 'Basic realm="${realm ?? 'angel_auth'}"'
-      ..end();
-    return false;
+      ..close();
   };
 }
 
-/// Use [requireAuthentication] instead.
-@deprecated
-final RequestMiddleware requireAuth = requireAuthentication(userKey: 'user');
-
 /// Restricts access to a resource via authentication.
-RequestMiddleware requireAuthentication({String userKey: 'user'}) {
+RequestHandler requireAuthentication<User>() {
   return (RequestContext req, ResponseContext res,
       {bool throwError: true}) async {
     bool _reject(ResponseContext res) {
@@ -31,7 +26,7 @@ RequestMiddleware requireAuthentication({String userKey: 'user'}) {
         return false;
     }
 
-    if (req.properties.containsKey(userKey) || req.method == 'OPTIONS')
+    if (req.container.has<User>() || req.method == 'OPTIONS')
       return true;
     else
       return _reject(res);